+++ src/JSContext.h
@@ -14,13 +14,12 @@ public:
 
     void    setGlobalValue(const std::string& name, JSValue value);
     bool    setFunction(JSFunctionIndex index, const std::string& source);
-    // bool evaluateBooleanFunction(JSFunctionIndex index);
 
 // protected:
     JSValue newNull();
 
     JSValue newBoolean(bool value);
-    JSValue newNumber(double value);
+    JSValue newNumber(float value);
     JSValue newString(const std::string& value);
     JSValue newArray();
     JSValue newObject();
@@ -31,11 +30,6 @@ public:
     void    resetToScopeMarker(JSScopeMarker marker);
 
 private:
-
-    // Used for proxy object.
-    // static int jsGetProperty(duk_context *_ctx);
-    // static int jsHasProperty(duk_context *_ctx);
-
     static void fatalErrorHandler(void* userData, const char* message);
 
     bool    evaluateFunction(uint32_t index);
@@ -43,5 +37,4 @@ private:
     JSValue getStackTopValue() { return JSValue(_ctx, duk_normalize_index(_ctx, -1)); }
 
     duk_context* _ctx = nullptr;
-    // friend JavaScriptScope<JSContext>;
 };
